resolve dtype errors and error with checkpoints

+++ main.py
@@ -29,7 +29,7 @@ def main():
     print("GPU devices:", torch.cuda.device_count())
 
     model_dtype = torch.float32
-    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=model_dtype).to(device)
+    model = AutoModelForCausalLM.from_pretrained(model_name, dtype=model_dtype).to(device)
     tokenizer = AutoTokenizer.from_pretrained(model_name)
     
     # --- 2. Initialize the quantizer ---
